+++ src/lib.rs
@@ -7,7 +7,6 @@
 //!
 //! ```no_run
 //! use linemux::MuxedLines;
-//! use tokio::stream::StreamExt;
 //!
 //! #[tokio::main]
 //! async fn main() -> std::io::Result<()> {
@@ -19,7 +18,7 @@
 //!
 //!     // Wait for `Line` event, which contains the line captured for a given
 //!     // source path.
-//!     while let Some(Ok(line)) = lines.next().await {
+//!     while let Ok(Some(line)) = lines.next_line().await {
 //!         println!("source: {}, line: {}", line.source().display(), line.line());
 //!     }
 //!     Ok(())